+++ packages/expo-router/src/react-navigation/routers/types.tsx
@@ -31,12 +31,6 @@ export type NavigationState<ParamList extends ParamListBase = ParamListBase> = R
    */
   routes: NavigationRoute<ParamList, keyof ParamList>[];
   /**
-   * Custom type for the state, whether it's for tab, stack, drawer etc.
-   * During rehydration, the state will be discarded if type doesn't match with router type.
-   * It can also be used to detect the type of the navigator we're dealing with.
-   */
-  type: string;
-  /**
    * Whether the navigation state has been rehydrated.
    */
   stale: false;
@@ -140,12 +134,6 @@ export type RouterConfigOptions = {
 };
 
 export type Router<State extends NavigationState, Action extends NavigationAction> = {
-  /**
-   * Type of the router. Should match the `type` property in state.
-   * If the type doesn't match, the state will be discarded during rehydration.
-   */
-  type: State['type'];
-
   /**
    * Initialize the navigation state.
    *